fix(chat): split the tool-activity pending provider from its hooks
`react(only-export-components)` failed the client lint: a module that
exports both a component and hooks breaks fast refresh. The provider now
lives in its own module and the context plus hooks stay in the hook file.
